Changed double -> float.

+++ Assets/Scripts/Data/CharacterData.cs
@@ -9,7 +9,10 @@ public static class CharacterData
 {
 	public static CharacterSelection CharacterSelection { get; set; }
 
-	public static double BloodSugarLevel { get; set; }
+	/// <summary>
+	/// Gets or sets the blood sugar level.
+	/// </summary>
+	public static float BloodSugarLevel { get; set; }
 }
 
 public enum CharacterSelection
